Refactor: TicketAddress package 변경
