+++ src/main/java/ticketaka/mtvs3_final_backend/member/command/application/service/MemberCommandService.java
@@ -8,8 +8,8 @@ import ticketaka.mtvs3_final_backend._core.error.exception.Exception400;
 import ticketaka.mtvs3_final_backend.member.command.application.dto.MemberCommandResponseDTO;
 import ticketaka.mtvs3_final_backend.member.command.domain.model.Address;
 import ticketaka.mtvs3_final_backend.member.command.domain.repository.AddressRepository;
-import ticketaka.mtvs3_final_backend.redis.ticketaddress.domain.TicketAddress;
-import ticketaka.mtvs3_final_backend.redis.ticketaddress.repository.TicketAddressRedisRepository;
+import ticketaka.mtvs3_final_backend.redis.ticket.address.domain.TicketAddress;
+import ticketaka.mtvs3_final_backend.redis.ticket.address.repository.TicketAddressRedisRepository;
 import ticketaka.mtvs3_final_backend.title.command.domain.model.Title;
 import ticketaka.mtvs3_final_backend.title.member.command.domain.model.MemberTitle;
 import ticketaka.mtvs3_final_backend.title.member.command.domain.repository.MemberTitleCommandRepository;
